metis-linker - organize & activate started data_frame ui
